refactor(web): atomize search routes into forms/handlers/responses (#665)
Split src/web/routes/search.py into the route-layering convention:
- src/web/search/forms.py — extract_length / parse_channel_id parsing
- src/web/search/handlers.py — orchestration (worker proxy, browse/local/
  semantic search, translate); returns domain DTOs, no FastAPI responses
- src/web/search/responses.py — SearchTemplate/Redirect/Json + search_response

Router becomes thin. External contract (URLs, query params, SSE-less JSON
shapes, redirect codes, error texts) unchanged. _extract_length re-exported
from the router for back-compat imports.
