+++ src/web/search/__init__.py
@@ -1,0 +1,1 @@
+"""Search web domain: forms, handlers, responses."""
